+++ archive/src/main/kotlin/org/openrs2/archive/cache/ExportCommand.kt
@@ -2,7 +2,7 @@ package org.openrs2.archive.cache
 
 import com.github.ajalt.clikt.core.CliktCommand
 import com.github.ajalt.clikt.parameters.arguments.argument
-import com.github.ajalt.clikt.parameters.types.long
+import com.github.ajalt.clikt.parameters.types.int
 import com.github.ajalt.clikt.parameters.types.path
 import com.google.inject.Guice
 import kotlinx.coroutines.runBlocking
@@ -10,7 +10,7 @@ import org.openrs2.archive.ArchiveModule
 import org.openrs2.cache.DiskStore
 
 public class ExportCommand : CliktCommand(name = "export") {
-    private val id by argument().long()
+    private val id by argument().int()
     private val output by argument().path(
         mustExist = true,
         canBeFile = false,
